+++ src/test/java/test/org/fugerit/java/xml2json/TestXmlToJsonHandler.java
@@ -1,6 +1,7 @@
 package test.org.fugerit.java.xml2json;
 
 import java.io.File;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.io.Reader;
 
@@ -19,7 +20,11 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class TestXmlToJsonHandler {
 
-	private boolean testWorker( String testId ) throws IOException, ConfigException, XMLException {
+	/*
+	 * Test XML to JSON
+	 */
+	
+	private boolean testWorkerXmlToJson( String testId ) throws IOException, ConfigException, XMLException {
 		File outputFile = new File( "target/"+testId+".json" );
 		String path = "cl://sample/"+testId+".xml";
 		log.info( "test worker path : {}, output : {}", path, outputFile.getCanonicalPath() );
@@ -33,13 +38,48 @@ public class TestXmlToJsonHandler {
 	
 	@Test
 	public void test1json() throws ConfigException, IOException, XMLException {
-		Assert.assertTrue( this.testWorker( "doc_test_01" ) );
+		Assert.assertTrue( this.testWorkerXmlToJson( "doc_test_01" ) );
 	}
 	
 	@Test
 	public void test2json() throws ConfigException, IOException, XMLException {
-		Assert.assertTrue( this.testWorker( "minimal" ) );
+		Assert.assertTrue( this.testWorkerXmlToJson( "minimal" ) );
 	}
+	
+	/*
+	 * Test JSON to XML
+	 */
+	
+	private boolean testWorkerJsonToXml( String testId ) throws IOException, ConfigException, XMLException {
+		File outputFile = new File( "target/"+testId+".xml" );
+		String path = "cl://sample/"+testId+".json";
+		log.info( "test worker path : {}, output : {}", path, outputFile.getCanonicalPath() );
+		try ( Reader reader = StreamHelper.resolveReader( path );
+				FileWriter writer = new FileWriter(outputFile) ) {
+			XmlToJsonHandler handler = new XmlToJsonHandler();
+			handler.writerAsXml(reader, writer);
+		}
+		return outputFile.exists();
+	}
+	
+	@Test
+	public void test1xml() throws ConfigException, IOException, XMLException {
+		Assert.assertTrue( this.testWorkerJsonToXml( "doc_test_01" ) );
+	}
+	
+	@Test
+	public void test2xmlFail1() throws ConfigException, IOException, XMLException {
+		Assert.assertThrows( ConfigException.class , () -> this.testWorkerJsonToXml( "fail_01" ) );
+	}
+
+	@Test
+	public void test3xmlFail2() throws ConfigException, IOException, XMLException {
+		Assert.assertThrows( ConfigException.class , () -> this.testWorkerJsonToXml( "fail_02" ) );
+	}
+	
+	/*
+	 * Test XML to YAML
+	 */
 	
 	private boolean testWorkerYaml( String testId ) throws IOException, ConfigException, XMLException {
 		File outputFile = new File( "target/"+testId+".yaml" );