+++ CHANGELOG.md
@@ -11,4 +11,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [maven build ans sonar scan workflow](.github/workflows/build_maven_package.yml)
 - xml to json conversion (see [xml conversion](src/main/docs/xml_conversion.md))
+- json to xml conversion (see [xml conversion](src/main/docs/xml_conversion.md))
 - [yaml support](src/main/docs/yaml_support.md) documentation